delete record.sh stop_record.sh

+++ quality-scripts/get_acceleration/get_acceleration.sh
@@ -1,8 +1,16 @@
 #/bin/bash
 SCRIPT_ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-#source ${SCRIPT_ROOT}/../common/log_function.sh
 case $1 in
     start)
+        cyber_channel echo /pnc/carstatus > ${SCRIPT_ROOT}/carstatus &
+        pid1=$!
+        sleep 60
+        if [ ! -s carstatus ]
+        then
+            echo "请检查/pnc/carstatus是否有输出"
+            exit 3
+        fi
+        kill -TERM $pid1
         cat carstatus | grep acc:| awk -F ":" '{print $2}' | awk '{print $1}' >${SCRIPT_ROOT}/acceleration
         python get_max_acceleration.py >$0.log &
         pid=$!
